requirements.dev.txt
-r requirements.txt
pytest-aiohttp
pytest
proxy.py
pyjwt < 2.0.0